backend/requirements.txt
Django>=6.0.0
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
django-storages>=1.14.0
boto3>=1.34.0
retrying>=1.3.4
djangorestframework-simplejwt>=5.3.1
pywebpush>=2.0.0
